refactor(encounter): route transform sites through dungeon.Module (Phase 3, part 1)
Phase 3 of the coordinate-types refactor (rpg-api issue #471). This first
part replaces all entities.Position / dungeon.Position usages in the
encounter orchestrator and entity types with dungeon.AbsolutePosition,
threading dungeon.Module through every transform site that bridges
room-local cube coordinates to dungeon-absolute.

Changes:
- entities.Position struct deleted from entities/encounter.go
- DoorInfo.Position, InitiativeEntry.Position, EntityStateData.Position,
  MovementCompletedEvent.FinalPosition, MonsterTurnCompletedEvent.Movement
  all migrate to dungeon.AbsolutePosition
- orchestrator.Position type alias pivots from entities.Position to
  dungeon.AbsolutePosition; all consumers updated
- moduleFromDungeon helper builds *dungeon.Module from entities.Dungeon
- localToAbsoluteOrLocal helper is the single transform site that lifts
  LocalPosition into AbsolutePosition (with identity fallback)
- cubeToLocalPosition replaces cubeToPosition (returns LocalPosition)
- convertToWallInfo / buildEntityMap / syncMonsterPositionFromRoom take
  Module + roomID parameters; walls leave the orchestrator already absolute
- MoveCharacter / ExecuteAction / EndTurn / OpenDoor / StartCombat all
  construct a Module and route every position through it
- Hand-rolled "cubeY := -cubeX - cubeZ" 2D-vestige math removed from the
  three monster-placement sites in orchestrator.go
- monster_turns.executeSingleMonsterTurn takes Module + currentRoomID and
  translates room-local movement to absolute when populating Movement
- Handler-side proto Position is int32 — the int(...) casts at the handler
  boundary replace the float64 widening that bridged the float-int gap
- Test fixtures updated: float64 expectations → int; entities.Position
  literals → dungeon.AbsolutePosition; mock setups for the new dungeon
  repo lookups in EndTurn

Acceptance progress (Phase 3):
- entities.Position type deleted
- All five transform sites routed through Module.LocalToAbsolute
- All tests pass: go test ./...

Remaining for Phase 3 part 2 (next commit):
- Delete dungeon.Position type from types.go
- Remove shiftRoomToAbsolute / shiftWallsByOrigin in handler converters
- Update docs (status.md, dungeon-component.md, encounter-handler.md,
  encounter-orchestrator.md, data-model.md)

Refs: rpg-project/ideas/coordinate-types/plan.md §Phase 3

+++ internal/handlers/dnd5e/v1alpha1/encounter/converters.go
@@ -445,11 +445,13 @@ func convertCombatStateToProto(
 	if state.CombatStarted && !state.CombatEnded && len(state.TurnOrder) > 0 {
 		activeEntry := state.TurnOrder[state.ActiveIndex]
 
-		// Convert position from service layer using cube coordinates for hex grids
+		// Convert position from service layer using cube coordinates for hex grids.
+		// activeEntry.Position is dungeon.AbsolutePosition (int); spatial.Position
+		// uses float64 — widening conversion is lossless.
 		var position *apiv1alpha1.Position
 		if activeEntry.Position != nil {
 			position = convertPositionToProto(
-				spatial.Position{X: activeEntry.Position.X, Y: activeEntry.Position.Y},
+				spatial.Position{X: float64(activeEntry.Position.X), Y: float64(activeEntry.Position.Y)},
 				gridType,
 				hexOrientation,
 			)
@@ -903,8 +905,8 @@ func convertMonsterTurnResultToProto(
 	// Convert movement path using cube coordinates for hex grids
 	movementPath := make([]*apiv1alpha1.Position, len(result.Movement))
 	for i, pos := range result.Movement {
-		// Convert encounter.Position to spatial.Position for cube conversion
-		spatialPos := spatial.Position{X: pos.X, Y: pos.Y}
+		// pos is dungeon.AbsolutePosition (int); spatial.Position uses float64.
+		spatialPos := spatial.Position{X: float64(pos.X), Y: float64(pos.Y)}
 		movementPath[i] = convertPositionToProto(spatialPos, gridType, hexOrientation)
 	}
